Move implementation to header and make cpp a soft link, define s_dump_message
