+++ examples/C++/zhelpers.hpp
@@ -167,13 +167,25 @@ s_sendmore (zmq::socket_t & socket, const std::string & string) {
     return (rc);
 }
 using ustring = std::basic_string<unsigned char>;
-inline static bool
-s_is_text_data(const ustring& data) {
-    for (int i = 0; i < data.size(); ++i) {
-        if (data[i]< 32 || data[i] > 127)
-            return false;
-    }
-    return true;
+inline static bool s_is_text_data(const ustring &data) {
+  for (int i = 0; i < data.size(); ++i) {
+    if (data[i] < 32 || data[i] > 127)
+      return false;
+  }
+  return true;
+}
+
+template <typename Stream>
+inline static void s_dump_message(Stream &os, const ustring &body) {
+  bool is_text = s_is_text_data(body);
+  os << "[" << std::setfill('0') << std::setw(3) << body.size() << "]";
+  for (size_t char_nbr = 0; char_nbr < body.size(); char_nbr++) {
+    if (is_text)
+      os << (char)body[char_nbr];
+    else
+      os << std::setfill('0') << std::setw(2) << std::hex
+         << (unsigned int)body[char_nbr];
+  }
 }
 
 //  Receives all message parts from socket, prints neatly
@@ -192,19 +204,7 @@ s_dump (zmq::socket_t & socket)
         size_t size = message.size();
         ustring data(static_cast<unsigned char*>(message.data()), size);
 
-        bool is_text = s_is_text_data(data);
-
-        size_t char_nbr;
-        unsigned char byte;
-
-        std::cout << "[" << std::setfill('0') << std::setw(3) << size << "]";
-        for (char_nbr = 0; char_nbr < size; char_nbr++) {
-            if (is_text)
-                std::cout << (char)data [char_nbr];
-            else
-                std::cout << std::setfill('0') << std::setw(2)
-                   << std::hex << (unsigned int) data [char_nbr];
-        }
+        s_dump_message(std::cout, data);
         std::cout << std::endl;
 
         int more = 0;           //  Multipart detection
